Modified the mageis_timseries.py file to plot only the MagEIS HighRate data. This was used to identify quit times for PAD fitting.
find_psd.py is a file that loads in the MagEphem file from this day, and
plots B_sc/B_eq, and L and MLT. There is a test to look for times when
L shell and MLT are close to the times of the microburst. There are two other
times with which psd_fit.py will fit.

The psd_fit.py fits a A*sin(aloha)^n function to the PAD. Plots are generated
in the data folder and fit parameters annotated in the fit plot. The fit
parameters are also saved in the data folder.